feat: run the migrator from its settings tab again
