+++ frontend/src2/router.ts
@@ -82,11 +82,6 @@ const routes = [
 		component: () => import('./data_source/DataSourceTable.vue'),
 	},
 	{
-		path: '/migration',
-		name: 'V2Migration',
-		component: () => import('./migration/Migration.vue'),
-	},
-	{
 		path: '/data-store',
 		name: 'DataStoreList',
 		component: () => import('./data_store/DataStoreList.vue'),
